refactor: Extract sortObjectKeys to shared cache utility
- Create shared cache-utils.ts with sortObjectKeys function
- Update server-cached-client.ts to use shared utility
- Update heatmap route to use shared utility
- Add missing filter params to heatmap cache key (minRating, onlyClassics,
  onlyTallClimbs, holdsFilter)
